Read Me
CorpuScrape is a Python tool created to fill a gap in text analysis research. The inspiration to develop this tool came from struggling to find an adequate dataset for a research project involving sentiment analysis of news articles. Furthermore, there were no free or open-source tools available to gather a dataset of this nature. In a few words, CorpuScrape helps researchers gather a dataset of news articles from different news sources for a specific topic and date range.

Building CorpuScrape required standing on the shoulders of giants. The following third-party Python libraries, APIs, and tools were combined:
Newspaper: “Simplified python article discovery & extraction.” 
This library is key as it takes in a URL and spits back content, authors, keywords, images, publishing date, and more with just a few lines of code.
BeautifulSoup: “HTML/XML parser for quick-turnaround applications like screen-scraping.”
BeautifulSoup parses all of the HTML content of a website and makes it possible to search for specific tags and extract any text or URL present within them.
PhantomJS: “A scripted, headless browser used for automating web page interaction.”
This is GUI-less browser can render JavaScript generated content from a website and output the page source which is then parsed by BeautifulSoup for searchability. This is crucial because BeautifulSoup can’t render JavaScript content.
Selenium: “Selenium automates browsers.”
We use a package that has bindings to control Selenium using Python. Selenium automates browser actions; it tells PhantomJS to click a specific button, or go to the next page to load more search results.
EasyGUI-QT: “A module for simple and easy GUI programming in Python.”
In order to make CorpuScrape more user-friendly, EasyGUI-QT was used to put a face on this tool. This module was chosen over the regular EasyGUI because it included a date picker.